[refactor] update logger reference and directory
